Hippo Scheduling Bugfix
ran into a hiccup with floating point math where, if there was no future scheduled departures, the default "scheduled departure" value would render as 24.999... instead of 25 hrs in the future. This would then get trimmed to 24 when converted to an integer, causing the hippo_clear automation to not trigger appropriately.

As a result, changed some templates to floats instead of integers, added extra seconds so it would render 25+, added more default values in case of errors.