breaking changes and tweaks
Staying ahead of breaking changes to how template sensors work. Tweaked departure notification for a longer delay because the charger was still 'on' as the car needs to temporarily charge to estimate when it needs to schedule the start.

Tweaked christmas tree movement automation because when a triggered sensor went off, it turned off the tree, despite other sensors still registering movement.